Fixed name parsing in case of v tags.
